pyproject.toml
[project]
name = "obsigpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.33",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openai>=0.3.38",
    "python-dotenv>=1.1.0",
]
